+++ Tests/Functional/Mapper/FrontEndUserMapperTest.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Tests\Functional\Mapper;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
-use OliverKlee\Oelib\Exception\NotFoundException;
 use OliverKlee\Seminars\Mapper\FrontEndUserMapper;
-use OliverKlee\Seminars\Model\FrontEndUser;
 use OliverKlee\Seminars\Model\FrontEndUserGroup;
 use OliverKlee\Seminars\Model\Registration;
 
@@ -80,107 +78,5 @@ final class FrontEndUserMapperTest extends FunctionalTestCase
         $registration = $model->getRegistration();
         self::assertInstanceOf(Registration::class, $registration);
         self::assertSame(1, $registration->getUid());
-    }
-
-    // Tests concerning findByUserName
-
-    /**
-     * @test
-     */
-    public function findByUserNameForEmptyUserNameThrowsException(): void
-    {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage('$value must not be empty.');
-
-        // @phpstan-ignore-next-line We are explicitly testing for a contract violation here.
-        $this->subject->findByUserName('');
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithNameOfExistingUserReturnsFrontEndUserInstance(): void
-    {
-        $userName = 'max.doe';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => $userName]);
-
-        self::assertInstanceOf(
-            FrontEndUser::class,
-            $this->subject->findByUserName($userName)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithNameOfExistingUserReturnsModelWithThatUid(): void
-    {
-        $userName = 'max.doe';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => $userName]);
-        $uid = (int)$this->getDatabaseConnection()->lastInsertId();
-
-        self::assertSame(
-            $uid,
-            $this->subject->findByUserName($userName)->getUid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithUppercasedNameOfExistingLowercasedUserReturnsModelWithThatUid(): void
-    {
-        $userName = 'max.doe';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => $userName]);
-        $uid = (int)$this->getDatabaseConnection()->lastInsertId();
-
-        self::assertSame(
-            $uid,
-            $this->subject->findByUserName(strtoupper($userName))->getUid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithUppercasedNameOfExistingUppercasedUserReturnsModelWithThatUid(): void
-    {
-        $userName = 'MAX.DOE';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => $userName]);
-        $uid = (int)$this->getDatabaseConnection()->lastInsertId();
-
-        self::assertSame(
-            $uid,
-            $this->subject->findByUserName($userName)->getUid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithLowercasedNameOfExistingUppercasedUserReturnsModelWithThatUid(): void
-    {
-        $userName = 'max.doe';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => strtoupper($userName)]);
-        $uid = (int)$this->getDatabaseConnection()->lastInsertId();
-
-        self::assertSame(
-            $uid,
-            $this->subject->findByUserName($userName)->getUid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithNameOfNonExistentUserThrowsException(): void
-    {
-        $this->expectException(NotFoundException::class);
-        $this->expectExceptionMessage('No records found in the table "fe_users" matching: {"username":"max.doe"}');
-
-        $userName = 'max.doe';
-        $this->getDatabaseConnection()->insertArray('fe_users', ['username' => $userName, 'deleted' => 1]);
-
-        $this->subject->findByUserName($userName);
     }
 }